+++ src/Illuminate/Foundation/Application.php
@@ -29,7 +29,7 @@ class Application extends Container implements ApplicationContract, HttpKernelIn
      *
      * @var string
      */
-    const VERSION = '5.7.7';
+    const VERSION = '5.7.8';
 
     /**
      * The base path for the Laravel installation.
