+++ src/Illuminate/Hashing/BcryptHasher.php
@@ -62,6 +62,8 @@ class BcryptHasher extends AbstractHasher implements HasherContract
      * @param  string  $hashedValue
      * @param  array  $options
      * @return bool
+     *
+     * @throws \RuntimeException
      */
     public function check($value, $hashedValue, array $options = [])
     {
