+++ src/Illuminate/Support/Facades/Event.php
@@ -37,7 +37,7 @@ class Event extends Facade
      * Replace the bound instance with a fake during the given callable's execution.
      *
      * @param  callable  $callable
-     * @param  array|string  $eventsToFake
+     * @param  array  $eventsToFake
      * @return callable
      */
     public static function fakeFor(callable $callable, array $eventsToFake = [])
